Fix problems in comments found by Alexander Kobotov
Just taking a quick look on the new release I noticed some minor issues there, which probably caused by a generation script:
- If you explore diff for files (c/z)gbrfsx, cgbsvx, cgbsvxx, cheequb, clanhf, zheequb you could find that number of datatype descriptions in documentation were changed to incorrect one (DOUBLE PRECISION instead of COMPLEX, or COMPLEX*16 instead of just COMPLEX). 
+ cunbdb has CMPLX instead of COMPLEX for X12 parameter.
+ zggevx RWORK become REAL instead of correct DOUBLE PRECISION
- Number of files got following string "/ output)", which seems meaningless. Just grep for it.

+++ lapack/trunk/SRC/zggevx.f
@@ -296,7 +296,7 @@
 *>
 *> \param[out] RWORK
 *> \verbatim
-*>          RWORK is REAL array, dimension (lrwork)
+*>          RWORK is DOUBLE PRECISION array, dimension (lrwork)
 *>          lrwork must be at least max(1,6*N) if BALANC = 'S' or 'B',
 *>          and at least max(1,2*N) otherwise.
 *>          Real workspace.
